chore: fixed test issue
