+++ .prettierignore
@@ -11,4 +11,4 @@ docs/api-content/**/*.json
 
 # Troublesome files
 tsconfig.json
-src/components/IconMapper/dynamicFontAwesomeImports.js
+src/components/IconMapper/dynamicFontAwesomeImports.ts
