+++ src/components/IconMapper/dynamicFontAwesomeImports.ts
@@ -1,0 +1,51 @@
+import { faCubes } from "@fortawesome/free-solid-svg-icons";
+import { faTerminal } from "@fortawesome/free-solid-svg-icons";
+import { faPersonWalkingLuggage } from "@fortawesome/free-solid-svg-icons";
+import { faObjectGroup } from "@fortawesome/free-solid-svg-icons";
+import { faEnvelopeOpenText } from "@fortawesome/free-solid-svg-icons";
+import { faDatabase } from "@fortawesome/free-solid-svg-icons";
+import { faHdd } from "@fortawesome/free-solid-svg-icons";
+import { faCloudArrowDown } from "@fortawesome/free-solid-svg-icons";
+import { faNetworkWired } from "@fortawesome/free-solid-svg-icons";
+import { faServer } from "@fortawesome/free-solid-svg-icons";
+import { faUsers } from "@fortawesome/free-solid-svg-icons";
+import { faWarehouse } from "@fortawesome/free-solid-svg-icons";
+import { faFlagCheckered } from "@fortawesome/free-solid-svg-icons";
+import { faPalette } from "@fortawesome/free-solid-svg-icons";
+import { faBook } from "@fortawesome/free-solid-svg-icons";
+import { faBookmark } from "@fortawesome/free-solid-svg-icons";
+import { faGavel } from "@fortawesome/free-solid-svg-icons";
+import { faTowerObservation } from "@fortawesome/free-solid-svg-icons";
+import { faLock } from "@fortawesome/free-solid-svg-icons";
+import { faGears } from "@fortawesome/free-solid-svg-icons";
+import { faScrewdriverWrench } from "@fortawesome/free-solid-svg-icons";
+import { faEyeSlash } from "@fortawesome/free-solid-svg-icons";
+import { faShield } from "@fortawesome/free-solid-svg-icons";
+import { faMicrochip } from "@fortawesome/free-solid-svg-icons";
+
+export const fontAwesomeIcons = {
+  cubes: faCubes,
+  terminal: faTerminal,
+  "person-walking-luggage": faPersonWalkingLuggage,
+  "object-group": faObjectGroup,
+  "envelope-open-text": faEnvelopeOpenText,
+  database: faDatabase,
+  hdd: faHdd,
+  "cloud-arrow-down": faCloudArrowDown,
+  "network-wired": faNetworkWired,
+  server: faServer,
+  users: faUsers,
+  warehouse: faWarehouse,
+  "flag-checkered": faFlagCheckered,
+  palette: faPalette,
+  book: faBook,
+  bookmark: faBookmark,
+  gavel: faGavel,
+  "tower-observation": faTowerObservation,
+  lock: faLock,
+  gears: faGears,
+  "screwdriver-wrench": faScrewdriverWrench,
+  "eye-slash": faEyeSlash,
+  shield: faShield,
+  microchip: faMicrochip,
+};
